moved typescript to common files (types.d.ts), some ts tweaks, removing all anys from the ts codeBase + simplified logic with new util function
