+++ .env
@@ -1,0 +1,2 @@
+VITE_APP_API_KEY = secret_token
+VITE_APP_API_URL = https://training.nerdbord.io/api/v1/panini-creator/order
