+++ src/components/componentsTypes.d.ts
@@ -1,0 +1,7 @@
+type SplashScreenProps = {
+  shouldTransition: boolean;
+  defaultPos: boolean;
+  navTo: string;
+  title?: string;
+  actionDesc?: string;
+};
