+++ src/components/paniniCreator/FormCard.tsx
@@ -1,10 +1,5 @@
 import styles from "./Form.module.css";
 
-export type FormProps = {
-  title: string;
-  children?: React.ReactNode;
-};
-
 export default function FormCard(props: FormProps) {
   return (
     <article className={styles.form}>
