+++ src/components/paniniCreator/formSections/CheckboxSection.tsx
@@ -1,16 +1,12 @@
 import styles from "./FormSection.module.css";
-import { FormSectionProps } from "./FormSectionProps";
 import FormSectionTemplate from "./FormSectionTemplate";
 import SpecialOptions from "./formSection/SpecialOptions";
 
-type CheckboxSectionProps = Omit<FormSectionProps, "removable"> & {
-  isBoolean: boolean;
-};
 export default function CheckboxSection(props: CheckboxSectionProps) {
   return (
     <FormSectionTemplate title={props.title}>
       <div className={styles.optionsWrapper}>
-        <SpecialOptions type="checkbox" options={props.options} name={props.name} isBoolean={props.isBoolean} />
+        <SpecialOptions type="checkbox" options={props.options} name={props.name} isValBoolean={props.isValBoolean} />
       </div>
     </FormSectionTemplate>
   );
