--- src/components/paniniCreator/formSections/FormSectionProps.d.ts
@@ -1,14 +1,0 @@
-export type FormSectionProps =
-  | {
-      removable: true;
-      maxElements: number;
-      title: string;
-      name: string;
-      options: string[];
-    }
-  | {
-      removable: false;
-      title: string;
-      name: string;
-      options: string[];
-    };
