+++ src/components/paniniCreator/formSections/formSection/Removals.tsx
@@ -1,15 +1,6 @@
 import { useEffect, useMemo, useState } from "react";
 import styles from "./Removals.module.css";
 
-type RemovalsProps = {
-  isActive: boolean;
-  maxElements: number;
-  defaultVal: string;
-  formElementsValues: string[];
-  setFormElementsValues: React.Dispatch<React.SetStateAction<string[]>>;
-  toggleActive: () => void;
-};
-
 export default function Removals(props: RemovalsProps) {
   const [isAdditionPossible, setIsAdditionPossible] = useState(true);
   useEffect(() => {
