+++ src/components/paniniCreator/formSections/formSection/elements/elementsTypes.d.ts
@@ -1,0 +1,12 @@
+type SelectElementProps = {
+  name: string;
+  options: string[];
+  formElementsValues: string[];
+  setFormElementsValues: React.Dispatch<React.SetStateAction<string[]>>;
+  orderVal: number;
+  defaultVal?: string;
+};
+
+interface SwipeElementProps extends SelectElementProps {
+  children?: ReactNode;
+}
