--- src/components/paniniCreator/formSections/formSection/options/OptionProps.d.ts
@@ -1,4 +1,0 @@
-export type OptionProps = {
-  option: string;
-  index: number;
-};
