+++ src/components/paniniCreator/formSections/formSection/options/optionsTypes.d.ts
@@ -1,0 +1,12 @@
+type OptionProps = {
+  option: readonly string;
+  index: number;
+};
+
+interface NamedOptionProps extends OptionProps {
+  name: string;
+}
+
+interface CheckboxOptionProps extends NamedOptionProps {
+  isValBoolean: boolean;
+}
