+++ src/components/paniniCreator/paniniCreatorTypes.d.ts
@@ -1,0 +1,4 @@
+type FormProps = {
+  title: string;
+  children?: React.ReactNode;
+};
