+++ src/data/cheese.ts
@@ -1,1 +1,1 @@
-export const cheeseVariants = ["EDAM", "MOZZARELLA", "STRACIATELLA", "GOUDA", "PECORINO"];
+export const cheeseVariants = ["EDAM", "MOZZARELLA", "STRACIATELLA", "GOUDA", "PECORINO"] as const;
