+++ src/data/vegetable.ts
@@ -1,1 +1,11 @@
-export const vegetableVariant = ["SALAD", "TOMATO", "CUCUMBER", "ONION", "PICKLES", "PEPPER", "ASPARAGUS", "BEETROOT", "OBERGINE"]
+export const vegetableVariant = [
+  "SALAD",
+  "TOMATO",
+  "CUCUMBER",
+  "ONION",
+  "PICKLES",
+  "PEPPER",
+  "ASPARAGUS",
+  "BEETROOT",
+  "OBERGINE",
+] as const;
