+++ src/pages/SplashScreenLayout.tsx
@@ -1,14 +1,6 @@
 import SplashScreen from "../components/SplashScreen";
 import PaniniCreator from "./PaniniCreator";
 
-type SplashScreenLayoutProps = {
-  shouldTransition: boolean;
-  defaultPos: boolean;
-  navTo: string;
-  title?: string;
-  actionDesc?: string;
-};
-
 export default function SplashScreenLayout(props: SplashScreenLayoutProps) {
   return (
     <>
