+++ src/types.d.ts
@@ -2,3 +2,5 @@ declare module "*.css" {
   const content: { [className: string]: string };
   export default content;
 }
+
+type SetOrderData = (val: object) => void;
